cart ui changes

+++ ecom-backend/src/main/java/org/example/ecombackend/model/Cart/CartDTO.java
@@ -1,9 +1,10 @@
 package org.example.ecombackend.model.Cart;
 import org.example.ecombackend.model.CartItem.CartItemDTO;
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartDTO (
         Long id,
         List<CartItemDTO> items,
-        double totalPrice
+        BigDecimal totalPrice
 ) {}
